fix data bug
